Split alerting code into another file
